Set log-level debug for docker and direct logs to journald (#213)
This commit restarts docker daemon for e2e instances with the settings to
set log-level to debug and directl docker logs to journald as well.
This can be useful for debugging in general and if capd nodes get
deleted and there's not much information in capi controllers logs.
The name of docker container representing a node is the same as the
CAPI machine object corresponding to it and we can check logs for the
deleted node with the following command:
`journalctl -b CONTAINER_NAME=name --all`