add process for generating log text analyzers based on collectors present (#334)
* add process for generating log text analyzers based on namespaced log collectors present

if we collect logs for a namespace, we will now include analyzers for those namespaced logs.

this allows us to generate log analyzers dynamically based on the namespaced log collector present; this is useful in cases like 'flux-system' or 'etcadm-controller-system', where the logs will not be present if the customer is not using gitops or external etcd.
if they are using those features, then we'll be collecting logs for those namespace, and will then subsequently add the analyzers for the given logs in the namespace.

* move namespace names to constants package for easier reference

* update mocks to accomodate renaming of package to 'diagnostics'

* use static map instead of building from empty map