Fix cluster name used during upgrade for CAPI objects (#545)
This commit includes two changes:
* Post upgrade eks-a waits for ControlPlaneReady and ManagedEtcdReady conditions
on the cluster. This check in the UpgradeCluster method in cluster_manager.go is
done on the workload cluster. In case of single management cluster, the workload
cluster in this method ends up being the management cluster itself since that maintains
all the CAPI objects. So this check will happen always only for the management cluster
instead of the target cluster undergoing upgrade.
This commit fixes this by passing the target workload cluster's name to the methods
that wait for the conditions post-upgrade
* Passes in the exact target cluster's name to retrieve the EtcdadmCluster CR
from the management cluster. This is how the KCP and MD objects are retrieved too