Mount extra volume when necessary (#582)
All eksa commands run inside docker using the mr-tools image.

When creating management cluster, users only have to specify a spec file,
and the cli will mount tha path that contains the spec as a volume to be
used by the mr-tools image.

With the introduction of workload clusters, users can pass an extra flag (--kubeconfig)
that points to the kubeconfig of a management cluster.
If this kubeconfig is not in the same path as the spec then it is necessary to
mount an extra volume to make it availalbe for the cli.

Fix #581