Delete worker node group name in old CLI E2E tests
The CLI does not allow extra fields in the cluster configuration file.
Old versions of the CLI don't support names for worker node groups,
since that a new field. This means we need to remove that field in the
E2E tests before passing the cluster config file to the old CLI.

Moved helpers to delete paths from a yaml object to the internal api
package and added tests.