Fix nodes to delete calculation for empty name
The old cluster spec (coming from v0.6.1 or prior) doesn't have a
default name for the first worker node group, since the default is
only set at the cli level and not at the api server level. So when
retrieving the cluster object from the kube api server, the name
is empty.

This is just a quick fix for this specific bug. We should find a more
robust way of implementing these defaults (probably with a mutation
webhook) so we don't need to implement the same logic again and again.