Handle replace directives for modules with CRDs
When adding CRDs to envtest we depend on the path the source modules are
on disk. These paths obviously change when go.mod has a replace
directive. With this changes we are able to handle those replaces and we
feed envtest with the path of the replacing module instead of the
replaced one.