Move context aware validation out of api package (#1356)
Validations made at the api package level should be lean, fast, free of
dependencies and context unaware.

This moves the registry mirror cert
validation out of the api package because it needs network connection
(this is done in order to check if the cert is self signed). It also
adds unit tests since the original code was missing them.