Dynamic config of api structs in cluster Config
This introduces the concept of Cluster Config Manager and Manager Entry.

* The manager holds the configuration to parse a Config from yaml, set its
defaults and validate it. It allows to dynamically register all of that
config, oferring a public api to Parse, SetDefaults and Validate.
* The entry is just a Manager api object to allow registering mappings,
processors, defaults and validations all at once. When adding new
content to the Config, we will probably need to register all of those
for the new type/s. This just makes it a bit easier.

The package public api to parse, set defaults and validate will continue
to be the same. The cluster package has a package level manager that is
used from ParseConfig, SetConfigDefaults and ValidateConfig.

Parsing has already been moved to the new manager framework. Defaults
and validations will follow in the next PRs.

Now when the parsing method finds an unknown Kind in the yaml content,
it ignores it instead of returning an error. This, aside for following a
bit better the idea of registering types, allows us to add new providers
as we go, without breaking the existing parsing until the new provider
is registered in the default manager.