Improve missing hardwarefile flag error speed (#1415)
The 'hardwarefile' flag is used with the tinkerbell provider only. It
specifies a file containing additional Kubernetes manifests that are
required for bare metal provisioning.

This patch ensures we don't perform unnecessary actions if the flag
isn't specified. Specifically it prevents a bunch of network and disk IO
that can be defered until we establish were in a good arg state.