Syncing from main -> cloudstack (#1510)
* Implement tinkerbell manifest writer for hardware generation. (#1482)

The yaml writer satisfies the MachineWriter interface required by
Translate functions for translating a machine description into
tinkerbell custom resource manifests.

* Fix nil pointer issue in upgrade plan with cilium (#1318)

* [PR BOT] Update ATTRIBUTION.txt file (#1457)

* pass the required info as flags instead of env (#1501)

* Enable only docker e2e tests (#1503)

* Implement json MachineWriter for Tinkerbell hardware generation. (#1500)

Similar to the yaml writer, we need to write tinkerbell hardware json
that can be registered with the tinkerbell stack. This cl adds the
construct for creating individual tinkerbell hardware jsons and
provides a mechanism to retrieve all files written enabling a separation
of generate and register steps.

Co-authored-by: Chris Doherty <chris.doherty4@gmail.com>
Co-authored-by: Taylor Neyland <57606775+taneyland@users.noreply.github.com>
Co-authored-by: EKS Distro PR Bot <75336432+eks-distro-pr-bot@users.noreply.github.com>
Co-authored-by: Abhinav Pandey <abhinavmpandey08@gmail.com>
Co-authored-by: Daniel Budris <budris@amazon.com>