Rework IP uniquness checks.
vSphere and Tinkerbell providers perform an IP uniqueness soft check by
attempting to connect to various services on a host. The algorithm was
attached to an IP generator forcing vSphere and Tinkerbell providers to
construct an IP generator before validating the IP.

The change ensures they can operate without needing an IP generator and
explicitly declares the check as 'soft'.