Rework ip uniqueness checks (#1543)
* Rework IP uniquness checks.

vSphere and Tinkerbell providers perform an IP uniqueness soft check by
attempting to connect to various services on a host. The algorithm was
attached to an IP generator forcing vSphere and Tinkerbell providers to
construct an IP generator before validating the IP.

The change ensures they can operate without needing an IP generator and
explicitly declares the check as 'soft'.

* Rename IsIPUnique to IsIPInUse.

The IsIPUnique check is inappropriately named as its not an authority on
whether or not an IP is assigned.

* Invert IsIPInUse() logic

The IsIPInUse() function that replaces IsIPUnique() has a name that
inverts the logic so should return true if in use and false if not.