Refactor release and eksd manifests reading
This extracts most of the logic to read and parse the different
manifests containing eksa, release and eksd CRDs from the cluster.Spec
into small reusable functions.

Created new packages, separating almost by CRD type. The dependency
between the new packages follows the same direction the different
manifests follow.

This is not an attempt to fully refactor cluster.Spec. There are still
problems with how it is constructed and the amount of functionality it has
amassed over time. Those problems will need to be tackled independently,
the only goal here was to just extract the CRD functionality.