Docker image mover (#1518)
Allows to set a source and destination for the images and orchestrates
the whole flow. For now, it admits disk and source and destination the
original registry as source and an external registry as destination.

Ideally we would clean the images loaded by the source after loading
them in the destination if they weren't there in the first place. We
will implement this as an improvement on top of this first iteration.
The drawback of not doing that cleanup is that we leave images in the
local docker storage, taking space.