Pass force flag to Tinkerbell provider (#1661)
* Fix required tinkerbell ip flag.

During hardware generation we require a tinkerbell IP unless the
--skip-registration flag is specified. This patches the behavior that
made the --tinkerbell-ip flag mandatory in all cases.

* Pass force flag to Tinkerbell provider

We want the force flag to create behavior that ensures machines are
usable when creating a new cluster even if they appear to have been
provisioned previously.

We can't model the behavior using dynamic construction logic so we're
passing the force flag all the way into the provider.