Process images concurrently in docker image mover
Little performance improvement for the docker image mover.

Now it runs image push and pulls concurrently. Before this is was doing
it sequentially, one after the other. It will use GOMAXPROCS/2 routines,
which should be equivalent to half the number of processor cores (in
most platforms).

Introduced a new matcher for contexts. It only checks that the argument
can be assigned to a `context.Context` variable.