Helm registry mirror
Instead of modifying the downloaded bundle manifests to point the helm
charts to the registry mirror, we leave them untouched, and we make helm
patch the requests urls on the fly.

The main advantage is that downloaded artifacts remain agnostic from the
registry. This allows to download them before having access to the
registry or even used the same artifacts for different registries.