Patch Tinkerbell workflow deletion. (#1727)
We don't separate static and interactive validation in Tinkerbell.
Consequently, the static yaml validation and interactive workflow
validation happen in the same function. This means we don't have an
opportunity post yaml validation to clean up workflows before we
validate none exist.

This change passes the force flag all the way down to validation logic
so it can warn in stead of error when it sees workflows alreaedy in the
Tinkerbell stack. In doing so, we gain yaml validation but have to
assume something will clean up the workflows (the scrub function) later
in the call path.