Download helm charts when saving images to disk (#1751)
This add support for helm charts to the download images command. It
creates a tarball with the tar that docker cli creates plus one tarball
per helm chart.

The command name 'download images' is not very accurate anymore and we
should rethink it. `download artifacts` would make a lot of sense but
since we already have an existing command using that pattern, I will
leave that refactor for a separate future PR.