Fix bug in GzipFileDownloadExtract
When using FileInfo.Name() we only have access to the file name and not
the complete path inside the tarball. This means that when having a
subfolder with the same name as the target file, the router would try to
unpackage it to the target path. Using the tar header name instead.