Remove wait for kubeconfig and RunPostControlPlaneCreation (#1848)
The wait for kubeconfig generation was added to run RunPostControlPlaneCreation,
which used to apply tolerations based on taints to the vsphere cloud controller
manager. Starting CAPV 1.0.1, the cloud controller daemonset is a part of cluster
template-cp.yaml, and eks-a applies required tolerations through this spec.
So we had removed the logic to apply tolerations from RunPostControlPlaneCreation.
This commit removes that interface method entirely as its no longer used in any providers
It removes the wait for kubeconfig secret generation as its unnecessary