Remove --setup-tinkerbell and --skip-power-actions (#2323)
For GA, --setup-tinkerbell is enforced as we don't support additional
cluster creation. We anticipate being able to detect Tinkerbell Enabled
Management Clusters and suspect this won't ever be required. If it is,
we can add it back in which is easier than removing it.

--skip-power-actions is superfluous as power actions are implied when a
user provides BMC configuration in the hardware CSV.