Build cluster.Config from a kubernetes client (#2359)
This introduces the ConfigClientBuilder that allows to register
processors to build different parts of the cluster.Config by retrieving
objects from a kubernetes cluster through a client. This is a very
similar pattern as the one used by the config manager.

Adds a default ConfigClientBuilder that pre-registers default
processors. For now, only identity providers, gitops and the snow
providers are implemented.