Add kubernetes.Client controller implementation
The controller runtime client implementation takes client.Client as an
input, which is a superset of runtime.Object. To make things easier, we
just make that a requirement in our kube client interface. To avoid
polluting our interface with something defined in controller runtime,
we just create a our own kubernetes.Object interface.