Packages: First end-to-end test (#2432) (#2448)
* Packages: First end-to-end test (#2432)

The first of what will eventually be many end-to-end tests for curated
packages. This test will:

* Bring up a cluster
* Install the curated packages controller (via helm)
* Install the hello-eks-anywhere package (via CLI)
* Forward a port to the hello-eks-anywhere service
* Verify the function of the hello-eks-anywhere package
* Tear down the cluster

There's plenty of room for future work, but some highlights include:

* Modify the Executable framework to allow the usage of exec.Command.Start()
and reading Stdout and Stderr before the process exits.

* Find a better way (than polling) to determine when a service is ready.

https://github.com/aws/eks-anywhere-packages/issues/96

* Packages: e2e test fixes for build pipelines

The port-forwarding was failing because the pod wasn't yet ready, so a simple
wait for the pod is added.

https://github.com/aws/eks-anywhere-packages/issues/96