Fix tinkerbell disk extractor
The disk extractor is used to extract disks from hardware so it can be
used to populate tinkerbell actions. The matching algorithm prevented
matches on hardware with labels other than those specified on the
selector which is inconsistent with how other code matches machines to
selectors.

The change uses the same matching function as all other parts of the
code base.