Avoid processing Snow images if the URI is empty (#2486)
It seems like the production Bundles is being generated without the Snow
sub bundle. This includes the images for kubevip and the manager. When
running "import images", the code retrieves the list of all images from
the Bundles and it fails because these two have an empty URI. This patch
avoids adding the snow images ot the list of the URI is empty.

We should remove this once the Bundles is fixed and the Snow components
are populated.