Add option to use native binaries in dep factory (#2683)
* Add option to use native binaries in dep factory

Until now, the executable builder always decided if docker needed to be
used or not, based on an env var. This made the constructor take all the
arguments, even when some of them aren't needed. Moreover, in order to
build local native executables programmatically through the factory, we
needed to set an env var and force the use of a "fake" tools image.

Now the factory has an specific option to use docker or not. If it's not
set, it will default to the value of the env var, keeping current
behavior. But it allows to override it.

* Fix typo in factory method name

* Fixed typo in doc comment

Co-authored-by: Joey Wang <jiayiwang7@yahoo.com>

* Fixed typo in doc comment

* Nest executables fields in factory

* Remove unnecessary tests

Co-authored-by: Joey Wang <jiayiwang7@yahoo.com>