Build provider cluster reconcilers at program init
Before this, provider cluster reconcilers were built dynamically on each
reconciliation loop. This required injecting all secondary dependencies
in the main cluster reconciler so they could be passed down to the
provider reconcilers when needed. This doesn't really scale with multiple
providers and made development difficult.

Instead, now we build the necessary providers during program
initialization. We check which are the CAPI providers that are installed
and we build the appropriate provider reconcilers based on that.

I made some minor refactors that are somehow not related with this
improvement but made working with this code easier and in a way shorten
the length of this PR. This includes: getting rid of the "abstract"
provider cluster reconciler and drying out some tests.