Include documentation on Ubuntu cgroups v2 (#2829)
The EKS Anywhere docs currently state that only Ubuntu 20.04.2 is
supported as the administration machine due to Ubuntu 21 and 22 using
cgroups v2. This commit explains how to configure Ubuntu 21 and 22 to
use cgroups v1 if the administration machine is using the latest Ubuntu
versions.

This has been tested using `make serve` from the docs path.