Use kubernetes client-go in a few places
In general, the EKS-Anywhere project has shelled out to kubectl to interact
with the cluster's API. However, given that the EKS-A kubectl binary is run
via a docker image, it can be slow. For cluster creation/teardown, this
slowness is amortized and isn't nearly as noticeable. But for commands like
"list packages" which should be rather fast, it's very noticeable.

So we're starting to use API client libraries where it makes sense to speed up
these calls that are more interactive in nature, or where faster responses
make sense.

There will be more commits of this nature in the near future. So if you see
patterns here that you don't like, or think could be improved, now would be a
great time to bring them up for discussion. :)

As a historical note, the choice to shell out to kubectl was made because it
was planned that EKS-A would support more versions of kubernetes than the
upstream project would, and it was thought that the ability to swap out
executable versions to match the kubernetes version would be an easier or more
flexible way to manage the differences when compared with linking to / using
multiple versions of the client libraries. Whether or not that is true remains
to be seen, but that fact is being used as guidance here, in that—at least at
first—we will replace only calls that are stable and therefore less likely to
change or require multiple library versions. (Like the list and get calls
replaced here).