Specify vsphere template in latest release tests (#3153)
Ubuntu doesn't support autoimport anymore so we can't rely on it to set
the template when creating clusters using the latest minor release.
Bottlerocket doesn't strictly need this but making the same change as
well for consistency. Plus it's a more robust solution since it doesn't
need to write to vsphere, avoiding race conditions with other test runs.

Using env vars for consistency with the rest of the input data. The env
var name is constructed from the OS and EKS-D version. These env vars
need to be set prior to running the tests.

When we release a new eks-a version with new eks-d versions, we need to
import template for those and add new variables for the tests to work
properly.