Improved validation of kubeconfig files
Instead of just checking that the file exists, the k8s client-go library is
used to load the file's data, which validates its basic structure, as well as
its existence.

This small step is a precursor to further work to concentrate how kubeconfig
files are handled across the various eksctl-anywhere subcommands.