fixup! more stuff from reviews
Ripping out the trashy stuff. There seemed to be consensus that we don't need
to mock the kubeconfig loader in order to test it.

However, I did verify that it errored on an empty file, and it turns out that
it doesn't... Which seems odd, but there you go. Since our previous
validations did error out on an empty file, I maintained that behavior, though
I doubt that it matters a ton either way.