Improved validation of kubeconfig files (#3189)
* Improved validation of kubeconfig files

Instead of just checking that the file exists, the k8s client-go library is
used to load the file's data, which validates its basic structure, as well as
its existence.

This small step is a precursor to further work to concentrate how kubeconfig
files are handled across the various eksctl-anywhere subcommands.

* fixup! Improved validation of kubeconfig files

* fixup! more stuff from reviews

Ripping out the trashy stuff. There seemed to be consensus that we don't need
to mock the kubeconfig loader in order to test it.

However, I did verify that it errored on an empty file, and it turns out that
it doesn't... Which seems odd, but there you go. Since our previous
validations did error out on an empty file, I maintained that behavior, though
I doubt that it matters a ton either way.