kubeconfig: Modify FromEnvironment to handle lists
The KUBECONFIG environment variable is defined to contain a list of
filenames. If it is a list, return the first one. If it's not a list, returns
the entire string (as before).

In theory, a program using this environment variable would merge the
kubeconfig files listed into a larger structure containing contexts, users,
and clusters. At this time however, the eksctl-anywhere command is always
expecting a single user/context/cluster to be defined in a single kubeconfig
file.