refactor(CLI): DRYs up kubeconfig checking in packages commands
This bit of repetitive code was lifted into the kubeconfig package, where it
can be modified once to prevent someone from having to do more shotgun commits
like this in the future.

The behavior for kubeconfig flags remains the same as previously in
aws/eks-anywhere#3304 and aws/eks-anywhere#3299.

Prow, why do you forsake me?