feat(CLI): Add ValidateFileOrEnv function in kubeconfig package
ValidateFileOrEnv validates a given kubeconfig filename, or, if and only if
that filename is an empty string, falls back to a kubeconfig filename from the
environment, and validates it.

This should be useful for CLI commands that want to find the first specified
kubeconfig file that's fit to use.

While this isn't used yet, my next PR will use it. :)