feat(CLI): Add --kubeconfig flag to curated packages commands (#3304)
Previously kubeconfig information was only loaded from an environment
file. Now each of the curated packages commands takes an optional --kubeconfig
flag, that if specified, loads the kubeconfig from that file, and errors if
the file doesn't exist or fails to load. When no --kubeconfig is specified, it
uses the previous behavior of trying to load a kubeconfig from the KUBECONFIG
environment variable.

Flags have also been reorganized to be consistent with the style introduced in
aws/eks-anywhere#3299.