refactor(CLI): Use the new kubeconfig functions in CP CLI commands (#3463)
Refactors the curated packages CLI commands to use the new functions
introduced in ac5db7f. The non curated packages commands were not refactored,
because I'm not 100% sure it's safe to do so. I'll leave that exercise for a
later PR.

This helps simplify our k8s go client creation for CP CLI commands, which will
be coming in a follow-up commit.