refactor: Replace errors.As with errors.Is
In some cases, errors.As was being used instead of errors.Is. The primary
difference being that errors.As alters the target error. This can make for
awkward error initializations, but also costs a bit of time to process.

More to the point, tests were failing with specific versions of go (namely
1.19).

Docs:

https://pkg.go.dev/errors#As

https://pkg.go.dev/errors#Is