Remove conditional handling of nutanix assetconfig
Uncommenting `release-0.11` case for `TestGenerateBundleManifest`
to make tests pass. Test has `DevRelease` set to true for both
`main` and `release-0.11` branch. As such, the test tries to fetch
a git tag for a project that doesn’t exist in that `release-0.11`
branch and it fails. The reason this works in the case of `snow`
is because the project is part of the `release-0.11` branch.