Add NutanixBundle type to VersionsBundle (#3292)
* Add NutanixBundle type to VersionsBundle

NutanixBundle contains the cluster API controller image and version.

* Fix org for CAPX project source

was incorrectly set as kubernetes-sigs.

* Linter cleanup

* Update the main release bundle

Ran `make update-bundle-golden-files`.

* Handle nutanix asset and bundle for 0.11 release

Skip nutanix bundle and asset handling for 0.11 release.

* Remove eksdrelease tag for nutanix artifacts

This can be moved out as Nutanix bundle does not have separate artifacts
for each Kubernetes version.

* Remove handling for nutanixCloudProvider artifacts

This is unused code as currently there is no CCM image as part
of the bundle.

* Fix incorrect key for bundleImageArtifact

Correct key is `cluster-api-provider-nutanix`.

* Remove conditional handling of nutanix assetconfig

Uncommenting `release-0.11` case for `TestGenerateBundleManifest`
to make tests pass. Test has `DevRelease` set to true for both
`main` and `release-0.11` branch. As such, the test tries to fetch
a git tag for a project that doesn’t exist in that `release-0.11`
branch and it fails. The reason this works in the case of `snow`
is because the project is part of the `release-0.11` branch.

* Regenerate main-bundle-release.yaml

* Remove release config as a parameter for `GetBundleReleaseAssetsConfigMap`