feat(CLI): BundleClient: go-native client for bundle operations
The bundle client uses a go-native kubernetes client to communicate with a
cluster to perform operations relating to curated packages bundles.

The client is not yet actually used, that will be coming in a separate
commit/PR. I've split them up because of their size. This will work as a sort
of expand-contract process[1], where new features like the bundle client are
added, then existing code will be modified to use these new features, then the
old solutions for these operations will be removed. Hopefully this will keep
the PR sizes manageable, without leaving the main branch in a broken state.

[1]: https://martinfowler.com/bliki/ParallelChange.html