feat(CLI): go-native client for CLI commands
This adds functions for creating a go-native kubernetes client to be used by
CLI commands.

The client is not yet used, that will be coming in follow-up commits/PRs. I've
split them up because of their size. This will work as an expand-contract
process[1], where new features like this kubernetes client are added, then
existing code will be modified to use these new features, then the old
solutions for these operations will be removed. Hopefully this will keep the
PR sizes manageable, without leaving the main branch in a broken state.

[1]: https://martinfowler.com/bliki/ParallelChange.html