Make WorkerNodeConfiguration.Count A Pointer (#3624)
* Change WorkerNodeGroupConfiguration count int->*int

* Add ptr package to get int pointer values

* Update tests with simple refs to Count

* Alter test expected error message

This test is supposed to exercise a flow where the cluster upgrade fails.
It does not matter what it fails on, so we can change the expected message without a problem.

* Change business logic to use *int

* Set node count default to 1 if nil

* Validate Count not nill in validation

This block should never fire, but it's better to have it in place in case a bug is introduced elsewhere.

* Add tests for coverage