Enable linters: cyclo comp, docs and naked returns
* revive: replacement for the deprecated golint. Enabled 3 issues that
are disabled by default for exported member comments and file/dir
permissions.

* godot: checks that exported member comments start with upper case and
end with a period (".").

* nakedret: forbids naked returns.

* gocyclo: measures cyclomatic complexity of functions. Restricted to a
max of 10. Recommended is 10-20. This requisite can be lowered down if
we find it to restrictive.

Now the linters only run in the diff between the current code and main.
This means it will only trigger issues for infringent new code. We
should strive to fix all old ocurrences if possible and add exceptions
for the rest. so we can renable a full repo scan.