Ensure Additional Trust bundle is propagated to the Nutanix Client (#3771)
* Ensure Additional Trust bundle is propagated to the Nutanix Client

The additional trust bundle was not being injected into the Nutanix Prism client.
This PR ensures the certificate provided in the additional trust bundle is tacked
on as a trusted certificate when constructing the Nutanix client.

* Fix import ordering

Appeasing the linter.

* Add additional tests for covering failure scenarios

Add tests for invalid values in additionalTrustBundle

* Appeasing the linter

run fixes for `make lint` issues.

* Move nutanix testdata files into a separate directory

Move nutanix testdata files to testdata/nutanix.

* Update nutanix pre-flight validation logic

Add certificate validation in nutanix pre-flight checks.