Remove vsphere provider dependency on api structs
It makes the code easier to follow, given the cluster spec is passed to
most methods as an argument. This reduces api structs copies to one,
removing ambiguity around which one contains the most updated data. It
also avoids bugs where updates are only performed in one place but not
the other.

If needed, a stateful version of the provider can be implemented by
composing on top of this one.