Update CP infra machine template names if changes (#3900)
When changes are detected in the provider machine templates for the
control plane or etcdadm cluster, we generate a new name because these
objects are immutable. If that happens, we also need to update their
corresponding references in the KubeadmControlPlane and EtcdadmCluster
or new machines will be never be rolled out.