Add custom logic to reconcile a CAPI control plane (#3993)
When using a control plane with external etcd, upgrades can't be done by
just applying al changes, and depending on the changes, it needs to be
orchestrated through annotations.

Marking the etcdadm cluster as upgrading, signals the kcp to not react
to changes in the etcd endpoints until the annotation is removed. This
will be done by the etcdadm cluster controller once the rolling upgrade
is finished.

Pausing the kcp reconciliation avoids having a unnecessary rolling
upgrade of the control plane when there are both cp and etcd changes
(for example, a k8s version upgrade). The CAPI cluster controller will
remove the pause annotation once the etcdadm cluster becomes ready.