Stop reconciliation if child objects are invalid
If datacenter or machine configs are invalid for a cluster,
reconciliation should stop to prevent any further changes.

When the failure message is not set, it's safe for us to assume the
object hasn't been validated yet but still might be valid. In this case,
we shouldn't bubble an error to the cluster status and just return and
wait for the object to updated, either to be valid or to have a failure
message.