Improve controller manager logs (#4115)
This now makes use of the tooling provided by upstream
controller-runtime and component-base/logs. As a consequence, the
manager now takes a standard set of flags to configure logging.

The default logging format is configured to json, it's the direction
upstream is going duwe to its tooling support. For tilt, using text
for now until we provide better tooling to analyze the logs.

Instead of injecting a logger in the reconciler on construction, we now
use contextual logging. The benefit here is controller-runtime will
inject a default set of values that uniquely identify the reconcile
request.