Fix delete workload clusters from CLI
The new full lifecycle controller adds a finalizer to eks-a Clusters to
be able to orchestrate its deletion. This interferes with the delete CLI
operation, since it doesn't yet make use of the controller and it
implements its own logic. The CLI tried to delete the eks-a cluster and
get blocked forever since the controller wouldn't remove the finalizer
for a paused cluster.

To solve this, we introduce a new annotation that signals the controller
that the cluster is being managed the CLI and hence it should allow to
be deleted without going through its normal deletion process.

There were also other hidden problems in the way cluster manager
implemented the deletion, by wrapping a bunch of API operations in one
big retry block. Some of these are not idempotent in combination with
the others (for example, try to update annotations after deleting a
resource). In addition, some of these operations had their own retry
mechanisms internally, causing the total timeout to increase
significantly. Now each individual operation has its own retry
mechanism.